Temporary solution - remove 'N' from sequence
When invalid symbol N is encountered, this no longer results in an
error. Instead, the N is removed from the sequence. Right now, the gap
is just removed which results in incorrect contexts defined by the
follow-up symbols. Therefore, in the future we want to split the
sequence and restart model construction / performance calculation after
an invalid symbol is encountered. Furthermore, support for other invalid
sequences (besides N) should be included.